Docs: make Markdown negotiation work
The proxy sat at the package root, but with an src/app layout Next only loads
src/proxy.ts, so it never ran. Its patterns were also built from
docsRoute = "/" as "/{/*path}", which matches nothing. Move it, fix the
patterns, skip the /llms.* routes, and handle the index page.

Agents that fetch pages with Accept: text/markdown or a .md suffix now get the
Markdown source with all code blocks.
